Updated UI look and added UI with buttons for ability items
I used Seth's prefab of the UI and adapted it to incorporate the colours of the character models in game. I added UI that related to the rigged dice and stun items with buttons for later programming. This was done in a separate scene and the UI was not overridden.